+++ projects/netgrif-application-engine/src/lib/data-fields/filter-field/models/filter-field.ts
@@ -10,8 +10,8 @@ export class FilterField extends DataField<string> {
     constructor(stringId: string, title: string,
                 initialValue: string, private _filterMetadata: FilterMetadata, private _allowedNets: Array<string>,
                 behavior: Behavior, placeholder: string, description: string, layout?: Layout,
-                validations?: Array<Validation>, component?: Component) {
-        super(stringId, title, initialValue, behavior, placeholder, description, layout, validations, component);
+                validations?: Array<Validation>, component?: Component, parentTaskId?: string) {
+        super(stringId, title, initialValue, behavior, placeholder, description, layout, validations, component, parentTaskId);
     }
 
     get filterMetadata(): FilterMetadata {