+++ projects/netgrif-application-engine/src/lib/data-fields/i18n-field/i18n-text-field/abstract-i18n-text-field.component.spec.ts
@@ -11,12 +11,12 @@ import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {AbstractI18nTextFieldComponent} from './abstract-i18n-text-field.component';
 import {AuthenticationMethodService} from '../../../authentication/services/authentication-method.service';
 import {MockAuthenticationMethodService} from '../../../utility/tests/mocks/mock-authentication-method-service';
-import {MockUserResourceService} from '../../../utility/tests/mocks/mock-user-resource.service';
-import {UserResourceService} from '../../../resources/engine-endpoint/user-resource.service';
-import {MockAuthenticationService} from '../../../utility/tests/mocks/mock-authentication.service';
-import {TestConfigurationService} from '../../../utility/tests/test-config';
-import {ConfigurationService} from '../../../configuration/configuration.service';
 import {AuthenticationService} from '../../../authentication/services/authentication/authentication.service';
+import {MockAuthenticationService} from '../../../utility/tests/mocks/mock-authentication.service';
+import {UserResourceService} from '../../../resources/engine-endpoint/user-resource.service';
+import {MockUserResourceService} from '../../../utility/tests/mocks/mock-user-resource.service';
+import {ConfigurationService} from '../../../configuration/configuration.service';
+import {TestConfigurationService} from '../../../utility/tests/test-config';
 import {TranslateService} from '@ngx-translate/core';
 
 describe('AbstractI18nTextFieldComponent', () => {