+++ projects/netgrif-application-engine/src/lib/event/model/event-outcomes/data-outcomes/get-data-localised-event-outcome.ts
@@ -1,0 +1,8 @@
+import {LocalisedFields} from '../../../../resources/interface/fields';
+import {TaskEventOutcome} from '../task-outcomes/task-event-outcome';
+
+export interface GetDataLocalisedEventOutcome extends TaskEventOutcome {
+
+    data: Array<LocalisedFields>;
+
+}