+++ projects/netgrif-application-engine/src/lib/task-content/model/resource-interface.ts
@@ -37,4 +37,5 @@ export interface DataFieldResource {
     };
     allowedNets?: Array<string>;
     filterMetadata?: FilterMetadata;
+    parentTaskId?: string;
 }