+++ projects/netgrif-application-engine/src/lib/task-content/services/task-content.service.ts
@@ -7,12 +7,12 @@ import {SnackBarService} from '../../snack-bar/services/snack-bar.service';
 import {TranslateService} from '@ngx-translate/core';
 import {EnumerationField, EnumerationFieldValue} from '../../data-fields/enumeration-field/models/enumeration-field';
 import {MultichoiceField} from '../../data-fields/multichoice-field/models/multichoice-field';
-import {ChangedFields, FrontendActions} from '../../data-fields/models/changed-fields';
+import {Change, ChangedFields, FrontendActions} from '../../data-fields/models/changed-fields';
 import {FieldConverterService} from './field-converter.service';
-import {EventOutcome} from '../../resources/interface/event-outcome';
 import {FieldTypeResource} from '../model/field-type-resource';
 import {DynamicEnumerationField} from '../../data-fields/enumeration-field/models/dynamic-enumeration-field';
 import {Validation} from '../../data-fields/models/validation';
+import {TaskEventOutcome} from '../../event/model/event-outcomes/task-outcomes/task-event-outcome';
 import {DataField} from '../../data-fields/models/abstract-data-field';
 
 /**
@@ -34,6 +34,12 @@ export abstract class TaskContentService implements OnDestroy {
     protected _task: Task;
     protected _taskDataReloadRequest$: Subject<FrontendActions>;
     protected _isExpanding$: BehaviorSubject<boolean>;
+    private _taskFieldsIndex: {
+        [taskId: string]: {
+            [fieldId: string]: DataField<any>;
+        }
+    } = {};
+    private _referencedTaskAndCaseIds: { [caseId: string]: Array<string> } = {};
 
     protected constructor(protected _fieldConverterService: FieldConverterService,
                           protected _snackBarService: SnackBarService,
@@ -77,6 +83,23 @@ export abstract class TaskContentService implements OnDestroy {
      */
     public get taskDataReloadRequest$(): Observable<FrontendActions> {
         return this._taskDataReloadRequest$.asObservable();
+    }
+
+
+    get taskFieldsIndex(): { [p: string]: { [p: string]: DataField<any> } } {
+        return this._taskFieldsIndex;
+    }
+
+    set taskFieldsIndex(value: { [p: string]: { [p: string]: DataField<any> } }) {
+        this._taskFieldsIndex = value;
+    }
+
+    get referencedTaskAndCaseIds(): { [p: string]: Array<string> } {
+        return this._referencedTaskAndCaseIds;
+    }
+
+    set referencedTaskAndCaseIds(value: { [p: string]: Array<string> }) {
+        this._referencedTaskAndCaseIds = value;
     }
 
     /**
@@ -141,7 +164,7 @@ export abstract class TaskContentService implements OnDestroy {
         if (!this._task || !this._task.dataGroups) {
             return false;
         }
-        const exists = this._task.dataGroups.some(group => group.fields.some( field => field instanceof DynamicEnumerationField ));
+        const exists = this._task.dataGroups.some(group => group.fields.some(field => field instanceof DynamicEnumerationField));
         if (!exists) {
             return true;
         }
@@ -183,11 +206,11 @@ export abstract class TaskContentService implements OnDestroy {
     /**
      * Clears the assignee, start date and finish date from the managed Task.
      */
-    public updateStateData(eventOutcome: EventOutcome): void {
+    public updateStateData(eventOutcome: TaskEventOutcome): void {
         if (this._task) {
-            this._task.user = eventOutcome.assignee;
-            this._task.startDate = eventOutcome.startDate;
-            this._task.finishDate = eventOutcome.finishDate;
+            this._task.user = eventOutcome.task.user;
+            this._task.startDate = eventOutcome.task.startDate;
+            this._task.finishDate = eventOutcome.task.finishDate;
         }
     }
 
@@ -199,68 +222,75 @@ export abstract class TaskContentService implements OnDestroy {
         if (!this._task || !this._task.dataGroups) {
             return;
         }
+        // todo actions owner zbytočný?
+        const frontendActions = chFields.taskId === this.task.stringId && chFields[TaskContentService.FRONTEND_ACTIONS_KEY];
 
-        const frontendActions = chFields.frontendActionsOwner === this.task.stringId && chFields[TaskContentService.FRONTEND_ACTIONS_KEY];
-
-        for (const dataGroup of this._task.dataGroups) {
-            for (const field of dataGroup.fields) {
-                if (chFields[field.stringId]) {
-                    if (this._fieldConverterService.resolveType(field) === FieldTypeResource.TASK_REF) {
-                        this._taskDataReloadRequest$.next(frontendActions ? frontendActions : undefined);
-                        return;
-                    }
-
-                    const updatedField = chFields[field.stringId];
-                    Object.keys(updatedField).forEach(key => {
-                        switch (key) {
-                            case 'type':
-                                // type is just an information, not an update. A field cannot change its type
-                                return; // continue - the field does not need updating, since nothing changed
-                            case 'value':
-                                field.valueWithoutChange(this._fieldConverterService.formatValueFromBackend(field, updatedField[key]));
-                                break;
-                            case 'behavior':
-                                if (updatedField.behavior[this._task.transitionId]) {
-                                    // TODO NGSD-489 fix behavior resolution
-                                    field.behavior = updatedField.behavior[this._task.transitionId];
-                                } else {
-                                    // ignore the behavior update, since it does not affect this task
-                                    return; // continue - the field does not need updating, since nothing changed
-                                }
-                                break;
-                            case 'choices':
-                                const newChoices: Array<EnumerationFieldValue> = [];
-                                if (updatedField.choices instanceof Array) {
-                                    updatedField.choices.forEach(it => {
-                                        newChoices.push({key: it, value: it} as EnumerationFieldValue);
-                                    });
-                                } else {
-                                    Object.keys(updatedField.choices).forEach(choiceKey => {
-                                        newChoices.push({key: choiceKey, value: updatedField.choices[choiceKey]} as EnumerationFieldValue);
-                                    });
-                                }
-                                (field as EnumerationField | MultichoiceField).choices = newChoices;
-                                break;
-                            case 'options':
-                                const newOptions = [];
-                                Object.keys(updatedField.options).forEach(optionKey => {
-                                    newOptions.push({key: optionKey, value: updatedField.options[optionKey]});
-                                });
-                                (field as EnumerationField | MultichoiceField).choices = newOptions;
-                                break;
-                            case 'validations':
-                                field.replaceValidations(updatedField.validations.map(it => (it as Validation)));
-                                break;
-                            default:
-                                field[key] = updatedField[key];
-                        }
-                        field.update();
-                    });
-                }
+        Object.keys(chFields).forEach(changedField => {
+            if (!!this.taskFieldsIndex[chFields.taskId] && !!this.taskFieldsIndex[chFields.taskId][changedField]) {
+                this.updateField(chFields, this.taskFieldsIndex[chFields.taskId][changedField], frontendActions);
             }
-        }
+        });
+
         this.$shouldCreate.next(this._task.dataGroups);
         this.performFrontendAction(frontendActions);
+    }
+
+    private updateField(chFields: ChangedFields, field: DataField<any>, frontendActions: Change): void {
+        if (this._fieldConverterService.resolveType(field) === FieldTypeResource.TASK_REF) {
+            this._taskDataReloadRequest$.next(frontendActions ? frontendActions : undefined);
+            return;
+        }
+
+        const updatedField = chFields[field.stringId];
+        Object.keys(updatedField).forEach(key => {
+            switch (key) {
+                case 'type':
+                    // type is just an information, not an update. A field cannot change its type
+                    return; // continue - the field does not need updating, since nothing changed
+                case 'value':
+                    field.valueWithoutChange(this._fieldConverterService.formatValueFromBackend(field, updatedField[key]));
+                    break;
+                case 'behavior':
+                    if (updatedField.behavior[this._task.transitionId]) {
+                        // TODO NGSD-489 fix behavior resolution
+                        field.behavior = updatedField.behavior[this._task.transitionId];
+                    } else {
+                        // ignore the behavior update, since it does not affect this task
+                        return; // continue - the field does not need updating, since nothing changed
+                    }
+                    break;
+                case 'choices':
+                    const newChoices: Array<EnumerationFieldValue> = [];
+                    if (updatedField.choices instanceof Array) {
+                        updatedField.choices.forEach(it => {
+                            newChoices.push({key: it, value: it} as EnumerationFieldValue);
+                        });
+                    } else {
+                        Object.keys(updatedField.choices).forEach(choiceKey => {
+                            newChoices.push({
+                                key: choiceKey,
+                                value: updatedField.choices[choiceKey]
+                            } as EnumerationFieldValue);
+                        });
+                    }
+                    (field as EnumerationField | MultichoiceField).choices = newChoices;
+                    break;
+                case 'options':
+                    const newOptions = [];
+                    Object.keys(updatedField.options).forEach(optionKey => {
+                        newOptions.push({key: optionKey, value: updatedField.options[optionKey]});
+                    });
+                    (field as EnumerationField | MultichoiceField).choices = newOptions;
+                    break;
+                case 'validations':
+                    field.replaceValidations(updatedField.validations.map(it => (it as Validation)));
+                    break;
+                default:
+                    field[key] = updatedField[key];
+
+            }
+            field.update();
+        });
     }
 
     /**