+++ projects/netgrif-application-engine/src/lib/task/services/cancel-task.service.spec.ts
@@ -13,7 +13,7 @@ import {TaskRequestStateService} from './task-request-state.service';
 import {NAE_TASK_OPERATIONS} from '../models/task-operations-injection-token';
 import {SubjectTaskOperations} from '../models/subject-task-operations';
 import {UnlimitedTaskContentService} from '../../task-content/services/unlimited-task-content.service';
-import {MessageResource} from '../../resources/interface/message-resource';
+import {EventOutcomeMessageResource} from '../../resources/interface/message-resource';
 import {Observable, of, throwError} from 'rxjs';
 import {map} from 'rxjs/operators';
 import {TaskResourceService} from '../../resources/engine-endpoint/task-resource.service';
@@ -29,6 +29,10 @@ import {TaskEvent} from '../../task-content/model/task-event';
 import {MockAuthenticationMethodService} from '../../utility/tests/mocks/mock-authentication-method-service';
 import {TaskDataService} from './task-data.service';
 import {DataFocusPolicyService} from './data-focus-policy.service';
+import {TaskEventOutcome} from '../../event/model/event-outcomes/task-outcomes/task-event-outcome';
+import {createMockCase} from '../../utility/tests/utility/create-mock-case';
+import {createMockNet} from '../../utility/tests/utility/create-mock-net';
+import {ChangedFieldsService} from '../../changed-fields/services/changed-fields.service';
 
 describe('CancelTaskService', () => {
     let service: CancelTaskService;
@@ -52,6 +56,7 @@ describe('CancelTaskService', () => {
                 TaskRequestStateService,
                 TaskDataService,
                 DataFocusPolicyService,
+                ChangedFieldsService,
                 {provide: TaskContentService, useClass: UnlimitedTaskContentService},
                 {provide: NAE_TASK_OPERATIONS, useClass: SubjectTaskOperations},
                 {provide: ConfigurationService, useClass: TestConfigurationService},
@@ -81,7 +86,13 @@ describe('CancelTaskService', () => {
                 fullName: ''
             },
             roles: {
-                role: 'perform'
+                role: {
+                    assign: true,
+                    cancel: true,
+                    finish: true,
+                    set: true,
+                    view: true
+                }
             },
             startDate: [1],
             finishDate: [1],
@@ -92,6 +103,7 @@ describe('CancelTaskService', () => {
             layout: {rows: 1, cols: 1, offset: 0},
             dataGroups: [],
             users: {},
+            userRefs: {},
             _links: {}
         };
         TestBed.inject(TaskContentService).task = testTask;
@@ -108,7 +120,43 @@ describe('CancelTaskService', () => {
     it('should cancel successfully', done => {
         expect(testTask.startDate).toBeTruthy();
         expect(testTask.user).toBeTruthy();
-        resourceService.response = {success: 'success'};
+        resourceService.response = {
+            success: 'success',
+            outcome: {
+                outcomes: [],
+                message: '',
+                task: {
+                    caseId: '',
+                    transitionId: '',
+                    title: '',
+                    caseColor: '',
+                    caseTitle: '',
+                    user: null,
+                    roles: {
+                        role: {
+                            assign: true,
+                            cancel: true,
+                            finish: true,
+                            set: true,
+                            view: true
+                        }
+                    },
+                    startDate: null,
+                    finishDate: null,
+                    assignPolicy: AssignPolicy.manual,
+                    dataFocusPolicy: DataFocusPolicy.manual,
+                    finishPolicy: FinishPolicy.manual,
+                    stringId: 'taskId',
+                    layout: {rows: 1, cols: 1, offset: 0},
+                    dataGroups: [],
+                    _links: {},
+                    users: {},
+                    userRefs: {}
+                },
+                aCase: createMockCase(),
+                net: createMockNet()
+            } as TaskEventOutcome
+        };
 
         let taskEvent: TaskEventNotification;
         taskEventService.taskEventNotifications$.subscribe(event => {
@@ -132,7 +180,10 @@ describe('CancelTaskService', () => {
     it('should cancel unsuccessful', done => {
         expect(testTask.startDate).toBeTruthy();
         expect(testTask.user).toBeTruthy();
-        resourceService.response = {error: 'error'};
+        resourceService.response = {
+            error: 'error',
+            outcome: {}
+        };
 
         let taskEvent: TaskEventNotification;
         taskEventService.taskEventNotifications$.subscribe(event => {
@@ -184,9 +235,9 @@ describe('CancelTaskService', () => {
 
 class TestTaskResourceService {
 
-    public response: MessageResource;
+    public response: EventOutcomeMessageResource;
 
-    public cancelTask(): Observable<MessageResource> {
+    public cancelTask(): Observable<EventOutcomeMessageResource> {
         if (this.response.error === 'throw') {
             return of(this.response).pipe(map(r => {
                 throw throwError(r);
@@ -199,7 +250,10 @@ class TestTaskResourceService {
 class TestUserService {
     public get user() {
         return {
-            email: 'mail'
+            email: 'mail',
+            roles: [{
+                stringId: 'role'
+            }]
         };
     }