+++ projects/netgrif-application-engine/src/lib/view/case-view/service/case-view-service.ts
@@ -29,6 +29,9 @@ import {SearchIndexResolverService} from '../../../search/search-keyword-resolve
 import {AllowedNetsService} from '../../../allowed-nets/services/allowed-nets.service';
 import {SortableView} from '../../abstract/sortable-view';
 import {NewCaseCreationConfigurationData} from '../../../side-menu/content-components/new-case/model/new-case-injection-data';
+import {PermissionService} from '../../../authorization/permission/permission.service';
+import {EventOutcomeMessageResource} from '../../../resources/interface/message-resource';
+import {CreateCaseEventOutcome} from '../../../event/model/event-outcomes/case-outcomes/create-case-event-outcome';
 
 @Injectable()
 export class CaseViewService extends SortableView implements OnDestroy {
@@ -55,7 +58,8 @@ export class CaseViewService extends SortableView implements OnDestroy {
                 protected _processService: ProcessService,
                 resolver: SearchIndexResolverService,
                 @Optional() @Inject(NAE_NEW_CASE_COMPONENT) protected _newCaseComponent: any,
-                @Optional() @Inject(NAE_NEW_CASE_CONFIGURATION) newCaseConfig: NewCaseConfiguration) {
+                @Optional() @Inject(NAE_NEW_CASE_CONFIGURATION) newCaseConfig: NewCaseConfiguration,
+                protected _permissionService: PermissionService) {
         super(resolver);
         this._newCaseConfiguration = {...this.DEFAULT_NEW_CASE_CONFIGURATION};
         if (newCaseConfig !== null) {
@@ -212,11 +216,11 @@ export class CaseViewService extends SortableView implements OnDestroy {
                 title: null,
                 color: 'panel-primary-icon',
                 netId: nets[0].stringId
-            }).subscribe((response: Case) => {
+            }).subscribe((response: EventOutcomeMessageResource) => {
                 this._snackBarService.openSuccessSnackBar(this._translate.instant('side-menu.new-case.createCase')
                     + ' ' + this._translate.instant('side-menu.new-case.defaultCaseName'));
                 this.reload();
-                myCase.next(response);
+                myCase.next((response.outcome as CreateCaseEventOutcome).aCase);
                 myCase.complete();
             }, error => this._snackBarService.openErrorSnackBar(error.message ? error.message : error));
         });
@@ -226,13 +230,13 @@ export class CaseViewService extends SortableView implements OnDestroy {
     protected getNewCaseAllowedNets(): Observable<Array<PetriNetReferenceWithPermissions>> {
         if (this._newCaseConfiguration.useCachedProcesses) {
             return this._allowedNetsService.allowedNets$.pipe(
-                map(net => net.filter(n => this.canDo(PermissionType.CREATE, n)))
+                map(net => net.filter(n => this._permissionService.hasNetPermission(PermissionType.CREATE, n)))
             );
         } else {
             return this._allowedNetsService.allowedNets$.pipe(
                 switchMap(allowedNets => {
                     return this._processService.getNetReferences(allowedNets.map(net => net.identifier)).pipe(
-                        map(net => net.filter(n => this.canDo(PermissionType.CREATE, n)))
+                        map(net => net.filter(n => this._permissionService.hasNetPermission(PermissionType.CREATE, n)))
                     );
                 })
             );
@@ -290,38 +294,12 @@ export class CaseViewService extends SortableView implements OnDestroy {
         return this._user.hasAuthority(authority);
     }
 
-    public canDo(action: string, net: PetriNetReferenceWithPermissions): boolean {
-        if (!net
-            || !net.permissions
-            || !action
-            || !(net.permissions instanceof Object)
-        ) {
-            return false;
-        }
-        if (Object.keys(net.permissions).some(role =>
-            this._user.hasRoleById(role) ? net.permissions[role][action] === false : false)) {
-            return false;
-        }
-        if (!Object.keys(net.permissions).filter(role => Object.keys(net.permissions[role])
-            .some(perm => perm === action)).some(role =>
-            !!net.permissions[role][action])) {
-            return true;
-        }
-        return Object.keys(net.permissions).some(role =>
-            this._user.hasRoleById(role) ? !!net.permissions[role][action] : false
-        );
-    }
-
     /**
      * Determines whether the current user has the [VIEW]{@link PermissionType#VIEW} permission on the current case
      * @param aCase the tested case
      * @returns `true` if the current user has the `VIEW` permission on the tested case, `false` otherwise.
      */
     public viewEnabled(aCase: Case): boolean {
-        const user = this._user.user;
-        if (!!aCase.viewRoles && aCase.viewRoles.length > 0) {
-            return user.roles.some(role => aCase.viewRoles.includes(role.stringId));
-        }
-        return false;
+        return this._permissionService.hasCasePermission(aCase, PermissionType.VIEW);
     }
 }