+++ projects/netgrif-components/src/lib/navigation/group-navigation-component-resolver/default-components/default-tabbed-case-view/default-tabbed-case-view.component.spec.ts
@@ -40,7 +40,7 @@ describe('DefaultTabbedCaseViewComponent', () => {
                         navigationItemTaskData: [{fields: []}, {
                             fields: [
                                 new FilterField(
-                                    `60f7b6aef68fea27e706b067-${UserFilterConstants.FILTER_FIELD_ID}`,
+                                    `${UserFilterConstants.FILTER_FIELD_ID}`,
                                     '',
                                     '',
                                     {