+++ projects/netgrif-components/src/lib/panel/task-panel/task-panel.component.ts
@@ -22,7 +22,9 @@ import {
     TaskDataService,
     TaskEventService,
     TaskRequestStateService,
-    TaskViewService
+    TaskViewService,
+    PermissionService,
+    ChangedFieldsService
 } from '@netgrif/application-engine';
 import {TaskContentComponent} from '../../task-content/task-content/task-content.component';
 import {TranslateService} from '@ngx-translate/core';
@@ -44,6 +46,7 @@ import {CurrencyPipe} from '@angular/common';
         DataFocusPolicyService,
         AssignPolicyService,
         FinishPolicyService,
+        ChangedFieldsService,
         {provide: NAE_TASK_OPERATIONS, useClass: SubjectTaskOperations},
     ]
 })
@@ -67,10 +70,13 @@ export class TaskPanelComponent extends AbstractTaskPanelComponent {
                 @Optional() @Inject(NAE_TASK_PANEL_DISABLE_BUTTON_FUNCTIONS) protected _disableFunctions: DisableButtonFuntions,
                 @Optional() @Inject(BOOLEAN_VALUE_LABEL_ENABLED) protected isEnabled: boolean,
                 protected _parentInjector: Injector,
-                protected _currencyPipe: CurrencyPipe) {
+                protected _currencyPipe: CurrencyPipe,
+                protected _changedFieldsService: ChangedFieldsService,
+                protected _permissionService: PermissionService) {
         super(_taskContentService, _log, _taskViewService, _paperView, _taskEventService, _assignTaskService,
             _delegateTaskService, _cancelTaskService, _finishTaskService, _taskState, _taskDataService,
-            _assignPolicyService, _callChain, _taskOperations, _disableFunctions, _translate, _currencyPipe);
+            _assignPolicyService, _callChain, _taskOperations, _disableFunctions, _translate, _currencyPipe, _changedFieldsService,
+            _permissionService);
     }
 
     protected createContentPortal(): void {